+++ PSGSuite/Public/Users/New-GSUser.ps1
@@ -45,6 +45,11 @@ function New-GSUser {
 
     This parameter expects a 'Google.Apis.Admin.Directory.directory_v1.Data.UserExternalId[]' object type. You can create objects of this type easily by using the function 'Add-GSUserExternalId'
 
+    .PARAMETER Locations
+    The Location objects of the user
+
+    This parameter expects a 'Google.Apis.Admin.Directory.directory_v1.Data.UserLocation[]' object type. You can create objects of this type easily by using the function 'Add-GSUserLocation'
+
     .PARAMETER Organizations
     The organization objects of the user
 
@@ -146,6 +151,9 @@ function New-GSUser {
         [Google.Apis.Admin.Directory.directory_v1.Data.UserExternalId[]]
         $ExternalIds,
         [parameter(Mandatory = $false)]
+        [Google.Apis.Admin.Directory.directory_v1.Data.UserLocation[]]
+        $Locations,
+        [parameter(Mandatory = $false)]
         [Google.Apis.Admin.Directory.directory_v1.Data.UserOrganization[]]
         $Organizations,
         [parameter(Mandatory = $false)]
@@ -223,6 +231,18 @@ function New-GSUser {
                             $extIdList.Add($extId)
                         }
                         $body.ExternalIds = $extIdList
+                    }
+                    Locations {
+                        if ($null -ne $Locations) {
+                            $locationList = New-Object 'System.Collections.Generic.List`1[Google.Apis.Admin.Directory.directory_v1.Data.UserLocation]'
+                            foreach ($loc in $Locations) {
+                                $locationList.Add($loc)
+                            }
+                            $body.Locations = $locationList
+                        }
+                        else {
+                            $toClear['locations'] = $null
+                        }
                     }
                     Organizations {
                         $orgList = New-Object 'System.Collections.Generic.List`1[Google.Apis.Admin.Directory.directory_v1.Data.UserOrganization]'